move traj dataclasses to detebase
