+++ mdsuite/file_io/file_read.py
@@ -25,146 +25,9 @@ Summary
 -------
 """
 import abc
-import dataclasses
-import numpy as np
 import typing
 
-
-@dataclasses.dataclass
-class PropertyInfo:
-    """
-    Information of a trajectory property.
-    example:
-    pos_info = PropertyInfo('Positions', 3)
-    vel_info = PropertyInfo('Velocities', 3)
-
-    Attributes
-    ----------
-    name:
-        The name of the property
-    n_dims:
-        The dimensionality of the property
-    """
-    name: str
-    n_dims: int
-
-
-@dataclasses.dataclass
-class SpeciesInfo:
-    """
-    Information of a species
-
-    Attributes
-    ----------
-    name
-        Name of the species (e.g. 'Na')
-    n_particles
-        Number of particles of that species
-    properties: list of PropertyInfo
-        List of the properties that were recorded for the species
-    mass and charge are optional
-    """
-    name: str
-    n_particles: int
-    properties: list
-    mass: float = None
-    charge: float = 0
-
-    def __eq__(self, other):
-        same = self.name == other.name and \
-               self.n_particles == other.n_particles and \
-               self.mass == other.mass and \
-               self.charge == other.charge
-        if len(self.properties) != len(other.properties):
-            return False
-
-        for prop_s, prop_o in zip(self.properties, other.properties):
-            same = same and prop_s == prop_o
-        return same
-
-
-@dataclasses.dataclass
-class TrajectoryMetadata:
-    """
-    This metadata must be extracted from trajectory files to build the database into which the trajectory will be stored
-
-    Attributes
-    ----------
-    n_configurations:
-        Number of configurations of the whole trajectory.
-    species_list: list of SpeciesInfo
-        The information about all species in the system.
-    box_l: list of float
-        The simulation box size in three dimensions
-    sample_step: optional
-        The time between consecutive configurations.
-        E.g. for a simulation with time step 0.1 where the trajectory is written every 5 steps: sample_step = 0.5
-        Does not have to be specified (e.g. configurations from Monte Carlo scheme), but is needed for all dynamic observables.
-    temperatyre: optional
-        The set temperature of the system.
-        Optional because only applicable for MD simulations with thermostat. Needed for certain observables.
-    simulation_data: optional
-        All other simulation data that can be extracted from the trajectory metadata.
-        E.g. software version, pressure in NPT simulations, time step, ...
-
-    """
-    n_configurations: int
-    species_list: list
-    box_l: list
-    sample_step: float = None
-    temperature: float = None
-    simulation_data: dict = dataclasses.field(default_factory=dict)
-
-
-class TrajectoryChunkData:
-    """
-    Class to specify the data format for transfer from the file to the database
-    """
-    def __init__(self, species_list: list, chunk_size: int):
-        """
-
-        Parameters
-        ----------
-        species_list:
-            List of SpeciesInfo.
-            It contains the information which species are there and which properties are recoreded for each
-        chunk_size:
-            The number of configurations to be stored in this chunk
-        """
-        self.chunk_size = chunk_size
-        self.species_list = species_list
-        self._data = dict()
-        for sp_info in species_list:
-            self._data[sp_info.name] = dict()
-            for prop_info in sp_info.properties:
-                self._data[sp_info.name][prop_info.name] = np.zeros((chunk_size, sp_info.n_particles, prop_info.n_dims))
-
-    def add_data(self, data: np.ndarray, config_idx, species_name, property_name):
-        """
-        Add configuration data to the chunk
-        Parameters
-        ----------
-        data:
-            The data to be added, with shape (n_configs, n_particles, n_dims).
-            n_particles and n_dims relates to the species and the property that is being added
-        config_idx:
-            Start index of the configs that are being added.
-        species_name
-            Name of the species to which the data belongs
-        property_name
-            Name of the property being added
-
-        example:
-        Storing velocity Information for 42 Na atoms in the 17th iteration of a loop that reads 5 configs per loop:
-        add_data(vel_array, 16*5, 'Na', 'Velocities')
-        where vel.data.shape == (5,42,3)
-
-        """
-        n_configs = len(data)
-        self._data[species_name][property_name][config_idx:config_idx + n_configs, ...] = data
-
-    def get_data(self):
-        return self._data
+from mdsuite.database.simulation_database import TrajectoryChunkData, TrajectoryMetadata
 
 
 class FileProcessor(abc.ABC):
